Absorb export into presentation facade
