+++ AGENTS.md
@@ -97,7 +97,7 @@ This is a clean-architecture Rust workspace for performance budgets and baseline
 | `perfgate-api` | API models and authentication types for baseline service |
 | `perfgate-app` | Orchestration layer for CLI commands |
 | `perfgate-render` | Workspace-only compatibility wrapper for `perfgate::presentation::render` |
-| `perfgate-export` | Multi-format export (CSV, JSONL, HTML, Prometheus, JUnit) |
+| `perfgate-export` | Workspace-only compatibility wrapper for `perfgate::presentation::export` |
 | `perfgate-sensor` | Cockpit mode and sensor report generation |
 | `perfgate-github` | GitHub API and PR-comment integration |
 | `perfgate-domain::scaling` | Complexity and scaling analysis |
