+++ Cargo.toml
@@ -68,7 +68,6 @@ wiremock = "0.6"
 serde_json.workspace = true
 toml.workspace = true
 perfgate-types.workspace = true
-perfgate-export.workspace = true
 perfgate-sensor.workspace = true
 perfgate-error.workspace = true
 perfgate-domain.workspace = true
