+++ GEMINI.md
@@ -27,7 +27,7 @@ The architecture preserves SRP seams across the remaining workspace crates and a
 | `perfgate-api` | API models and authentication types for baseline service |
 | `perfgate-app` | Orchestration layer for CLI commands |
 | `perfgate-render` | Workspace-only compatibility wrapper for `perfgate::presentation::render` |
-| `perfgate-export` | Multi-format export (CSV, JSONL, HTML, Prometheus, JUnit) |
+| `perfgate-export` | Workspace-only compatibility wrapper for `perfgate::presentation::export` |
 | `perfgate-sensor` | Cockpit mode and sensor report generation |
 | `perfgate-github` | GitHub API and PR-comment integration |
 | `perfgate-domain::scaling` | Complexity and scaling analysis |
