+++ docs/WORKSPACE.md
@@ -7,7 +7,7 @@ This file is automatically generated by `cargo run -p xtask -- docs-sync`.
 | Crate | Description | Kill Rate Target |
 |-------|-------------|------------------|
 | `perfgate-error` | Compatibility wrapper for perfgate_types::error | 100% |
-| `perfgate-export` | Export formats (CSV, JSONL, HTML, Prometheus) | 90% |
+| `perfgate-export` | Workspace-only compatibility wrapper for perfgate::presentation::export | 90% |
 | `perfgate-render` | Workspace-only compatibility wrapper for perfgate::presentation::render | 90% |
 | `perfgate-sensor` | Sensor report builder for cockpit integration | 90% |
 | `perfgate-paired` | Paired benchmarking statistics (A/B testing) | 100% |
