+++ xtask/Cargo.toml
@@ -20,7 +20,7 @@ serde.workspace = true
 serde_json.workspace = true
 perfgate-types.workspace = true
 perfgate-domain.workspace = true
-perfgate-export.workspace = true
+perfgate-app.workspace = true
 glob.workspace = true
 regex.workspace = true
 toml.workspace = true
